Actions: Add `force_release` input

+++ update.sh
@@ -6,7 +6,7 @@ url='https://raw.githubusercontent.com/felixonmars/dnsmasq-china-list/master/acc
 new_list=$(mktemp)
 
 curl -sS --fail-early --fail-with-body "$url" | sed -nE 's!^server=/(([-a-z0-9]+\.)*[a-z]+)/[0-9.]+$!\1!p' | LC_ALL=C sort -u >$new_list
-if cmp -s $new_list default.list; then
+if [[ "${CHINA_PAC_FORCE_RELEASE:-}" != 'true' ]] && cmp -s $new_list default.list; then
     rm $new_list
     exit
 fi
